Add files via upload
We at LearnClue have come up with the next level revolutionary concept of making the Abacus accessible digitally through mobile app in Android and IOS. Our application suits best for the children of 6-14 years of age. In this application a child can access the Abacus virtually in a mobile app. The student will be guided by experts through scheduled regular live sessions.
It helps in solving typical mathematical calculations in very short time.It teaches time management, helps to solve problems in simple steps.It stands useful in competitive exams and talent tests.Our methodology encourages the learners to play games & engage in activities that can be used for learning the content, critical thinking, problem solving, decision making, for review & reinforcement.
The Government of India and the State Government recommend all the schools to introduce communication skills, public speaking, mind mapping and other special skills as part of their regular curriculum.The thought behind this initiative is to develop these skills in children right from the beginning, so that they would be well equipped to excel in their higher education and further in their jobs